requirements.txt
Django==4.2.20
psycopg2>=2.8
djangorestframework
djangorestframework-simplejwt
django-filter
redis[hiredis]
celery
